pass url to modifier run; add timestamp to logging; attach zndraw logger to socket
