+++ zndraw/app.py
@@ -136,6 +136,7 @@ def modifier_run(data):
         points=points,
         segments=segments,
         json_data=data["atoms"] if "atoms" in data else None,
+        url=data["url"],
     )
     io.emit("atoms:clear", int(data["step"]) + 1)
     for idx, atoms in tqdm.tqdm(enumerate(atoms_list)):
