+++ zndraw/utils.py
@@ -1,5 +1,5 @@
 import socket
-
+import logging
 
 def get_port(default: int = 1234) -> int:
     """Get an open port."""
@@ -14,3 +14,19 @@ def get_port(default: int = 1234) -> int:
     finally:
         sock.close()
     return port
+
+class ZnDrawLoggingHandler(logging.Handler):
+    """Logging handler which emits log messages to the ZnDraw server."""
+
+    def __init__(self, socket):
+        super().__init__()
+        self.socket = socket
+
+    def emit(self, record):
+        try:
+            msg = self.format(record)
+            self.socket.emit("message:log", msg)
+        except RecursionError:  # See StreamHandler
+            raise
+        except Exception:
+            self.handleError(record)
