+++ zndraw/view.py
@@ -2,6 +2,7 @@ import logging
 import webbrowser
 
 from zndraw.app import app, io
+from zndraw.utils import ZnDrawLoggingHandler
 
 try:
     import urllib.request
@@ -10,8 +11,11 @@ try:
 except ImportError:
     wv = None
 
-log = logging.getLogger("werkzeug")
-log.setLevel(logging.ERROR)
+werkzeug_log = logging.getLogger("werkzeug")
+werkzeug_log.setLevel(logging.ERROR)
+
+log = logging.getLogger("zndraw")
+log.setLevel(logging.INFO)
 
 
 def _view_with_webview(url, fullscreen=False):
@@ -62,4 +66,15 @@ def view(
         ).start()
     elif open_browser:
         webbrowser.open(url)
+    
+    logging_handler = ZnDrawLoggingHandler(io)
+    logging_handler.setLevel(logging.INFO)
+    # attach ISO timestamp to log messages
+    formatter = logging.Formatter(
+        "%(asctime)s.%(msecs)03d %(message)s", "%Y-%m-%dT%H:%M:%S"
+    )
+    logging_handler.setFormatter(formatter)
+
+    logging.getLogger("zndraw").addHandler(logging_handler)
+
     io.run(app, port=port, host="0.0.0.0")
