+++ zndraw/zndraw.py
@@ -17,24 +17,10 @@ import znh5md
 
 from zndraw.bonds import ASEComputeBonds
 from zndraw.data import atoms_from_json, atoms_to_json
-from zndraw.utils import get_port
+from zndraw.utils import get_port, ZnDrawLoggingHandler
 
 
-class ZnDrawLoggingHandler(logging.Handler):
-    """Logging handler which emits log messages to the ZnDraw server."""
 
-    def __init__(self, socket):
-        super().__init__()
-        self.socket = socket
-
-    def emit(self, record):
-        try:
-            msg = self.format(record)
-            self.socket.emit("message:log", msg)
-        except RecursionError:  # See StreamHandler
-            raise
-        except Exception:
-            self.handleError(record)
 
 
 @dataclasses.dataclass
